+++ src/lib/terra-components.module.ts
@@ -19,6 +19,8 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatDatepickerModule } from '@angular/material/datepicker';
+import { A11yModule } from '@angular/cdk/a11y';
+import { MatSelectModule } from '@angular/material/select';
 import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
@@ -39,8 +41,10 @@ import { MatInputModule } from '@angular/material/input';
         DragDropModule,
         MatIconModule,
         MatCheckboxModule,
-        MatFormFieldModule,
         MatDatepickerModule,
+        A11yModule,
+        MatFormFieldModule,
+        MatSelectModule,
         MatInputModule
     ],
     declarations: [...components, ...directives, ...pipes],